refactor MovieDetailsActivity a bit and add a spinner while fetching data
